refactor(compatibility-layer): explicitly showcase the difference between moment implementation and luxon implementation
