+++ packages/grafana-data/src/datetime/types.ts
@@ -1,0 +1,83 @@
+
+export interface DateTimeBuiltinFormat {
+  __momentBuiltinFormatBrand: any;
+}
+
+const ISO_8601_TOKEN = 'ISO_8601';
+
+export const ISO_8601: DateTimeBuiltinFormat = ISO_8601_TOKEN as unknown as DateTimeBuiltinFormat;
+export type DateTimeInput = Date | string | number | Array<string | number> | DateTime | null;
+export type FormatInput = string | DateTimeBuiltinFormat | undefined;
+export type DurationInput = string | number | DateTimeDuration;
+export type DurationUnit =
+  | 'year'
+  | 'years'
+  | 'y'
+  | 'month'
+  | 'months'
+  | 'M'
+  | 'week'
+  | 'weeks'
+  | 'isoWeek'
+  | 'w'
+  | 'day'
+  | 'days'
+  | 'd'
+  | 'hour'
+  | 'hours'
+  | 'h'
+  | 'minute'
+  | 'minutes'
+  | 'm'
+  | 'second'
+  | 'seconds'
+  | 's'
+  | 'millisecond'
+  | 'milliseconds'
+  | 'ms'
+  | 'quarter'
+  | 'quarters'
+  | 'Q';
+
+export interface DateTimeLocale {
+  firstDayOfWeek: () => number;
+}
+
+export interface DateTimeDuration {
+  asHours: () => number;
+  hours: () => number;
+  minutes: () => number;
+  seconds: () => number;
+  asSeconds: () => number;
+}
+
+export interface DateTime extends Object {
+  add: (amount?: DateTimeInput, unit?: DurationUnit) => DateTime;
+  clone: () => DateTime;
+  set: (unit: DurationUnit | 'date', amount: DateTimeInput) => void;
+  diff: (amount: DateTimeInput, unit?: DurationUnit, truncate?: boolean) => number;
+  endOf: (unitOfTime: DurationUnit) => DateTime;
+  format: (formatInput?: FormatInput) => string;
+  fromNow: (withoutSuffix?: boolean) => string;
+  from: (formaInput: DateTimeInput) => string;
+  isSame: (input?: DateTimeInput, granularity?: DurationUnit) => boolean;
+  isBefore: (input?: DateTimeInput) => boolean;
+  isValid: () => boolean;
+  local: () => DateTime;
+  locale: (locale: string) => DateTime;
+  startOf: (unitOfTime: DurationUnit) => DateTime;
+  subtract: (amount?: DateTimeInput, unit?: DurationUnit) => DateTime;
+  toDate: () => Date;
+  toISOString: (keepOffset?: boolean) => string;
+  isoWeekday: (day?: number | string) => number | string;
+  valueOf: () => number;
+  unix: () => number;
+  utc: () => DateTime;
+  utcOffset: () => number;
+  hour?: () => number;
+  minute?: () => number;
+  month?: () => number;
+  date?: () => number;
+  year?: () => number;
+  tz?: (timeZone?: string) => DateTime | string;
+}
